feat(astro-integration-angular): convert to typescript from javascript
chore: convert to typescript and test changes are working

chore: add types for component from @angular/core
