+++ packages/astro-integration-angular/package.json
@@ -4,9 +4,9 @@
   "description": "An Angular integration for Astro",
   "type": "module",
   "exports": {
-    ".": "./index.js",
-    "./client.js": "./client.js",
-    "./server.js": "./server.js",
+    ".": "./src/index.js",
+    "./client.js": "./src/client.js",
+    "./server.js": "./src/server.js",
     "./package.json": "./package.json"
   },
   "files": [
