+++ packages/astro-integration-angular/src/index.ts
@@ -1,6 +1,7 @@
-import angular from '@analogjs/vite-plugin-angular';
+import viteAngular from '@analogjs/vite-plugin-angular';
+import { AstroIntegration, AstroRenderer } from 'astro';
 
-function getRenderer() {
+function getRenderer(): AstroRenderer {
   return {
     name: '@analogjs/astro-integration-angular',
     clientEntrypoint: '@analogjs/astro-integration-angular/client.js',
@@ -14,11 +15,11 @@ function getViteConfiguration() {
       include: ['@analogjs/astro-integration-angular/client.js'],
       exclude: ['@analogjs/astro-integration-angular/server.js'],
     },
-    plugins: [angular.default()],
+    plugins: [(viteAngular as any).default()],
   };
 }
 
-export default function () {
+export default function (): AstroIntegration {
   return {
     name: '@analogjs/astro-integration-angular',
     hooks: {
